Added a victory sound, corrected a bug in volume management on SDL
git-svn-id: svn+ssh://src.fovea.cc/svn/flobopuyo/branches/autotools@2008 053fdb51-1f17-0410-be66-d00040120cd6

+++ fpcore/Story.cpp
@@ -205,7 +205,7 @@ static void music(StyrolyseClient *_this, const char *name)
     theCommander->playMusicTrack(name);
 }
 
-static void playSound(StyrolyseClient *_this, const char *fileName, int volume)
+static void playSound(StyrolyseClient *_this, const char *fileName, float volume)
 {
     theCommander->playSound(fileName, volume);
 }
